+++ preload.js
@@ -15,6 +15,15 @@ contextBridge.exposeInMainWorld('electronAPI', {
   getSessionsDir: () => ipcRenderer.invoke('get-sessions-dir'),
   getSessionIds: () => ipcRenderer.invoke('get-session-ids'),
   openSessionsFolder: () => ipcRenderer.invoke('open-sessions-folder'),
+  onKbEntry: (callback) => {
+    ipcRenderer.on('kb-entry', callback);
+  },
+  onKbPass: (callback) => {
+    ipcRenderer.on('kb-pass', callback);
+  },
+  onKbSl: (callback) => {
+    ipcRenderer.on('kb-sl', callback);
+  },
   onWindowPositionChanged: (callback) => {
     ipcRenderer.on('window-position-changed', (_event, pos) => callback(pos));
   }